--- packages/shared/react-utils.ts
@@ -1,7 +1,0 @@
-export interface IMutableRefObject {
-  current: any[]
-}
-
-export function isRef(v: any): v is IMutableRefObject {
-  return !!(v && Object.hasOwnProperty.call(v, 'current'))
-}
