+++ packages/shared/utils-common/index.ts
@@ -1,0 +1,3 @@
+export * from './array'
+export * from './common'
+export * from './number'
