+++ packages/shared/utils-common/number.ts
@@ -1,0 +1,1 @@
+export const isNumber = (val: unknown): val is number => typeof val === 'number'
