+++ packages/shared/utils-types/react/index.ts
@@ -1,0 +1,3 @@
+export interface IMutableRefObject {
+  current: any[]
+}
