+++ packages/shared/utils-types/vue/index.ts
@@ -1,0 +1,1 @@
+export { }
