+++ packages/shared/utils-vue/index.ts
@@ -1,0 +1,1 @@
+export { }
